requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
docstring_parser==0.16
frozenlist==1.4.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-discoveryengine==0.11.10
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.15
langchain-community==0.0.32
langchain-core==0.1.41
langchain-google-vertexai==0.1.2
langchain-text-splitters==0.0.1
langsmith==0.1.45
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.0
packaging==23.2
pandas==2.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
SQLAlchemy==2.0.29
tenacity==8.2.3
tqdm==4.66.2
types-protobuf==4.25.0.20240410
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
